Trying to use the verifier
